+++ encryption_playground/lib/features/home/presentation/constants/sidebar_dimensions.dart
@@ -1,0 +1,8 @@
+class SidebarDimensions {
+  static const double expanded = 260.0;
+  static const double compact = 72.0;
+
+  // No mobile, usually the drawer handles the width automatically or fills most of the screen.
+  // But we can define a standard drawer width if needed.
+  static const double mobileDrawer = 280.0;
+}
