+++ encryption_playground/lib/features/home/presentation/widgets/side_menu.dart
@@ -1,198 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:provider/provider.dart';
 
-import '../../../../../app/locale_controller.dart';
-import '../../../../../app/theme_controller.dart';
+import '../../../../../shared/responsive/responsive_breakpoints.dart';
 import '../../../../../shared/theme/app_colors.dart';
-import '../pages/home_page.dart';
+import '../constants/sidebar_dimensions.dart';
+import 'side_menu_compact.dart';
+import 'side_menu_expanded.dart';
 
 class SideMenu extends StatelessWidget {
   const SideMenu({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return ValueListenableBuilder<String>(
-        valueListenable: HomePageNavigationService.instance.routeTracker.currentRoute,
-        builder: (context, currentRoute, child){
-          return Drawer(
-            child: Container(
-              width: 250,
-              color: AppColors.sideMenuBackground,
-              child: Column(
-                children: [
+    final state = ResponsiveLayout.current(context);
+    final isCompact = state == ResponsiveLayoutState.compact;
+    final isMobile = state == ResponsiveLayoutState.mobile;
 
-                  // Logo
-                  Padding(
-                    padding: const EdgeInsets.all(16.0),
-                    child: Image.asset(
-                      'assets/logo_full.png',
-                      width: double.infinity,
-                      fit: BoxFit.contain,
-                    ),
-                  ),
-
-                  // Buttons
-                  Expanded(
-                    child: ListView(
-                      padding: const EdgeInsets.symmetric(horizontal: 16),
-                      children: [
-                        _SideMenuItem(
-                          title: 'Dashboard',
-                          icon: Icons.dashboard_outlined,
-                          isSelected: currentRoute == '/dashboard',
-                          onTap: ()=> HomePageNavigationService.instance.navigatorKey.currentState?.pushReplacementNamed('/dashboard'),
-                        ),
-                        const SizedBox(height: 8),
-                        _SideMenuItem(
-                          title: 'Ciphers',
-                          icon: Icons.security_outlined,
-                          isSelected: currentRoute == '/ciphers',
-                          onTap: () => HomePageNavigationService.instance.navigatorKey.currentState?.pushReplacementNamed('/ciphers'),
-                        ),
-                        const SizedBox(height: 8),
-                        _SideMenuItem(
-                          title: 'Symmetric',
-                          icon: Icons.sync_alt_outlined,
-                          isSelected: currentRoute == '/symmetric',
-                          onTap: () => HomePageNavigationService.instance.navigatorKey.currentState?.pushReplacementNamed('/symmetric'),
-                        ),
-                        const SizedBox(height: 8),
-                        _SideMenuItem(
-                          title: 'Asymmetric',
-                          icon: Icons.vpn_key_outlined,
-                          isSelected: currentRoute == '/asymmetric',
-                          onTap: () => HomePageNavigationService.instance.navigatorKey.currentState?.pushReplacementNamed('/asymmetric'),
-                        ),
-                        const SizedBox(height: 8),
-                        _SideMenuItem(
-                          title: 'Hash',
-                          icon: Icons.fingerprint_outlined,
-                          isSelected: currentRoute == '/hash',
-                          onTap: () => HomePageNavigationService.instance.navigatorKey.currentState?.pushReplacementNamed('/hash'),
-                        ),
-                      ],
-                    ),
-                  ),
-
-                  // Bottom section
-                  Padding(
-                    padding: const EdgeInsets.all(16.0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: [
-                        OutlinedButton.icon(
-                          onPressed: () {
-                            context.read<ThemeController>().toggleTheme();
-                          },
-                          icon: Icon(
-                            context.watch<ThemeController>().isDarkMode
-                                ? Icons.light_mode
-                                : Icons.dark_mode,
-                          ),
-                          label: const Text('Toggle Theme'),
-                        ),
-                        const SizedBox(height: 8),
-                        OutlinedButton.icon(
-                          onPressed: () {
-                            context.read<LocaleController>().toggleLocale();
-                          },
-                          icon: const Icon(Icons.language),
-                          label: Text(
-                            context.watch<LocaleController>().value.languageCode == 'en'
-                                ? 'English'
-                                : 'Português',
-                          ),
-                        ),
-                        const SizedBox(height: 16),
-                        Center(
-                          child: Text(
-                            'v1.0.0',
-                            style: GoogleFonts.spaceGrotesk(
-                              color: AppColors.sideMenuTextInactive,
-                              fontSize: 12,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        }
-    );
-  }
-}
-
-class _SideMenuItem extends StatefulWidget {
-  final String title;
-  final IconData icon;
-  final bool isSelected;
-  final VoidCallback onTap;
-
-  const _SideMenuItem({
-    required this.title,
-    required this.icon,
-    required this.isSelected,
-    required this.onTap,
-  });
-
-  @override
-  State<_SideMenuItem> createState() => _SideMenuItemState();
-}
-
-class _SideMenuItemState extends State<_SideMenuItem> {
-  bool _isHovered = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final color = widget.isSelected
-        ? AppColors.sideMenuTextActive
-        : AppColors.sideMenuTextInactive;
-    final bgColor = widget.isSelected
-        ? AppColors.sideMenuActiveBackground
-        : (_isHovered ? AppColors.sideMenuHoverBackground : Colors.transparent);
-
-    return MouseRegion(
-      onEnter: (_) => setState(() => _isHovered = true),
-      onExit: (_) => setState(() => _isHovered = false),
-      child: InkWell(
-        onTap: widget.onTap,
-        borderRadius: BorderRadius.circular(2),
-        child: AnimatedContainer(
-          duration: const Duration(milliseconds: 200),
-          curve: Curves.easeInOut,
-          decoration: BoxDecoration(
-            color: bgColor,
-            borderRadius: BorderRadius.circular(2),
-            border: Border(
-              left: BorderSide(
-                color: widget.isSelected ? AppColors.sideMenuTextActive : AppColors.sideMenuTextInactive,
-                width: 2,
-              ),
-            ),
-          ),
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-            child: Row(
-              children: [
-                Icon(widget.icon, color: color, size: 20),
-                const SizedBox(width: 12),
-                Text(
-                  widget.title,
-                  style: GoogleFonts.spaceGrotesk(
-                    color: color,
-                    fontWeight: widget.isSelected ? FontWeight.w700 : FontWeight.w500,
-                    fontSize: 14,
-                  ),
-                ),
-              ],
-            ),
-          ),
+    if (isMobile) {
+      return Drawer(
+        child: Container(
+          color: AppColors.sideMenuBackground,
+          child: const SideMenuExpanded(),
         ),
-      ),
+      );
+    }
+
+    return AnimatedContainer(
+      duration: const Duration(milliseconds: 250),
+      curve: Curves.easeInOut,
+      width: isCompact ? SidebarDimensions.compact : SidebarDimensions.expanded,
+      color: AppColors.sideMenuBackground,
+      child: isCompact ? const SideMenuCompact() : const SideMenuExpanded(),
     );
   }
 }
